refactor the out.py filename
